+++ src/main/java/com/appointments/appointments/coordinator/CoordinatorService.java
@@ -4,6 +4,7 @@ import com.appointments.appointments.appUser.AppUser;
 import com.appointments.appointments.appUser.AppUserService;
 import com.appointments.appointments.coordinator.dto.CoordinatorRequest;
 import com.appointments.appointments.coordinator.dto.CoordinatorResponse;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -24,7 +25,8 @@ public class CoordinatorService {
     // =========================================================================
 
     public CoordinatorResponse findById(Integer id){
-        Coordinator coordinator = coordinatorRepository.findById(id).orElse(new Coordinator());
+        Coordinator coordinator = coordinatorRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Coordinator Not Found"));
 
         return coordinatorMapper.toCoordinatorResponse(coordinator);
     }
@@ -39,7 +41,9 @@ public class CoordinatorService {
     public CoordinatorResponse updateCoordinator(Integer id ,CoordinatorRequest dto){
         AppUser appUser = appUserService.findAppUserByIdEntity(dto.appUserId());
 
-        Coordinator coordinator = coordinatorRepository.findById(id).orElse(new Coordinator());
+        Coordinator coordinator = coordinatorRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Coordinator Not Found"));
+
         coordinator.setName(dto.name());
 
         coordinator = coordinatorRepository.save(coordinator);
@@ -48,15 +52,23 @@ public class CoordinatorService {
     }
 
     public void deleteById(Integer id){
+        if(!coordinatorRepository.existsById(id)) throw new EntityNotFoundException("Coordinator Not Found");
+
         coordinatorRepository.deleteById(id);
     }
-
 
     // =========================================================================
     // METHODS FOR THE SERVICES (Retrieves Entities)
     // =========================================================================
 
+    public Coordinator findByIdEntity(Integer id){
+        return coordinatorRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Coordinator Not Found"));
+    }
+
     public Coordinator createCoordinatorEntity(Coordinator coordinator){
         return coordinatorRepository.save(coordinator);
     }
+
+
 }
