+++ src/main/java/com/appointments/appointments/pacient/PacientService.java
@@ -2,6 +2,7 @@ package com.appointments.appointments.pacient;
 
 import com.appointments.appointments.pacient.dto.PacientRequest;
 import com.appointments.appointments.pacient.dto.PacientResponse;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +30,8 @@ public class PacientService {
     }
 
     public PacientResponse findById(Integer id){
-        Pacient pacient = pacientRepository.findById(id).orElse(new Pacient());
+        Pacient pacient = pacientRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Pacient Not Found"));
 
         return pacientMapper.toPacientDto(pacient);
     }
@@ -42,7 +44,9 @@ public class PacientService {
     }
 
     public PacientResponse updatePacient(Integer id, PacientRequest dto){
-        Pacient pacient = pacientRepository.findById(id).orElse(new Pacient());
+        Pacient pacient = pacientRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Pacient Not Found"));
+
         pacient.setName(dto.name());
         pacient.setEmail(dto.email());
 
@@ -52,6 +56,8 @@ public class PacientService {
     }
 
     public void deleteById(Integer id){
+        if(!pacientRepository.existsById(id)) throw new EntityNotFoundException("Pacient Not Found");
+
         pacientRepository.deleteById(id);
     }
 
@@ -61,6 +67,11 @@ public class PacientService {
     // =========================================================================
 
     public Pacient findByIdEntity(Integer id){
-        return pacientRepository.findById(id).orElse(new Pacient());
+        return pacientRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Pacient Not Found"));
+    }
+
+    public Boolean existById(Integer id){
+        return pacientRepository.existsById(id);
     }
 }
